// features/citizens/createCitizen/types.ts
export default interface FormValues {
  firstName: string;
  lastName: string;
  middleName: string;
  suffix: string;
  birthdate: string;
  sex: '' | '0' | '1';
  district: string;
  inhabitantType: string;
  citizenType: string;
  citizenTypeClassification: string;
  email: string;
  mobileNumber: string;
  username: string;
}